// packages/client/src/game/lib/constants/assets/tiles.ts
export enum Tilesets {
  Resource = "resource",
  BoundsOuterBorder = "bounds-outerborder",
  BoundsInnerBorder = "bounds-innerborder",
  BoundsNonBuildable = "bounds-nonbuildable",
}

export enum Tilemaps {
  AsteroidMicro = "asteroid-micro",
  AsteroidSmall = "asteroid-small",
  AsteroidMedium = "asteroid-medium",
  AsteroidLarge = "asteroid-large",
}

export enum ResourceTilekeys {
  Copper = 1,
  Iron,
  Lithium,
  Sulfur,
  Titanium,
  Kimberlite,
  Iridium,
  Platinum,
}